Store multiple locations per friend, spiderify markers
This is a pretty substantial change to the Selenium parts
of the code, and processing urls without selenium will now
either break or produce a javascript variable incompatible
with map.html.

The good news is that selenium processing (now the default
option) will now fetch _all_ places associated with a persion,
ie where that person lived, worked, went to school, etc.

This required associated changes to geocoding and the map
javascript.

In addition, to make this new plethora of results more usable,
I added spiderification to the map pointers. Clicking on a
marker-dense area will expand the markers away from eachother with
a line tracing back to their original location. Two finger clicking
on a marker will open the associated facebook page.